Refactor TraceGraph structure
