+++ web/src/features/trace/components/TraceGraph/BasicEdge/index.ts
@@ -1,0 +1,1 @@
+export { BasicEdge } from "./BasicEdge";
