+++ web/src/features/trace/components/TraceGraph/BasicEdge/styles.ts
@@ -1,0 +1,26 @@
+export const styles = {
+  edgeLabelContainer: {
+    display: "flex",
+    justifyContent: "center",
+    flexDirection: "column",
+    background: "#1e1e1e",
+    textAlign: "center",
+    width: "fit-content",
+    height: "fit-content",
+    margin: "auto",
+    fontSize: 12,
+    "&:hover": {
+      cursor: "default",
+    },
+  },
+  timeContainer: {
+    color: "#ffffff",
+  },
+  counterContainer: {
+    background: "#3b3c42",
+    border: 1,
+    borderColor: "#96979e",
+    borderRadius: 0.5,
+    padding: 1,
+  },
+};
