+++ web/src/features/trace/components/TraceGraph/BasicNode/BasicNode.tsx
@@ -4,8 +4,8 @@ import { Handle, NodeProps, Position } from "reactflow";
 
 import { ResourceIcon } from "@/components/Elements/ResourceIcon";
 
+import { NodeData } from "../types";
 import { styles } from "./styles";
-import { NodeData } from "./types";
 
 const BasicNodeImpl = (props: NodeProps<NodeData>) => {
   const { image, name, type, color } = props.data;
@@ -17,16 +17,16 @@ const BasicNodeImpl = (props: NodeProps<NodeData>) => {
         position={Position.Top}
         style={{ visibility: "hidden" }}
       />
-      <Box sx={styles.nodeStyle.nodeBox}>
+      <Box sx={styles.nodeBox}>
         <Box
           sx={{
-            ...styles.nodeStyle.nodeIconBox,
+            ...styles.nodeIconBox,
             borderColor: color,
           }}
         >
           <ResourceIcon name={image} />
         </Box>
-        <Box sx={styles.nodeStyle.nodeText}>
+        <Box sx={styles.nodeText}>
           <Box sx={{ color: "#E9EAF1" }}>{name}</Box>
           <Box sx={{ fontSize: 14 }}>{type}</Box>
         </Box>
