+++ web/src/features/trace/components/TraceGraph/BasicNode/index.ts
@@ -1,0 +1,1 @@
+export { BasicNode } from "./BasicNode";
