+++ web/src/features/trace/components/TraceGraph/utils/global.ts
